requirements.txt
streamlit
pillow
colorthief